requirements.txt
python-dotenv>=0.21.0
python-binance>=1.0.16
